Requests as a lazy import dependency
