+++ gymnos/services/http.py
@@ -9,14 +9,71 @@ import uuid
 import shutil
 import logging
 
+from tqdm import tqdm
 from collections.abc import Iterable
 
 from .service import Service
-from ..utils.downloader import download_file_from_url
-from ..utils.text_utils import filenamify_url
 from ..utils.hashing import sha1_text
+from ..utils.text_utils import filenamify_url
+from ..utils.lazy_imports import lazy_imports as lazy
 
 logger = logging.getLogger(__name__)
+
+
+def urljoin(*args):
+    """
+    Joins given arguments into an url. Trailing but not leading slashes are
+    stripped for each argument.
+    """
+
+    return "/".join(s.strip("/") for s in args)
+
+
+def download_file_from_url(url, file_path, force=False, verbose=True, headers=None, raise_for_status=True,
+                           chunk_size=1024):
+    """
+    Download url to local file.
+
+    Parameters
+    ---------
+    url: str
+        Url to download
+    file_path: str
+        Path to download file
+    force: bool, optional
+        Whether or not force download. By default downloads are cached
+    verbose: bool, optional
+        Whether or not show progress bar
+    raise_for_status: bool, optional
+        Whether or not raise error if status code is not success
+    chunk_size: int, optional
+        Chunk size to read stream.
+
+    Returns
+    -------
+    requests.Response
+        Response object
+    """
+    if os.path.isfile(file_path) and not force:
+        return
+
+    response = lazy.requests.get(url, stream=True, headers=headers)
+    response.raise_for_status()
+
+    iterator = response.iter_content(chunk_size=chunk_size)
+
+    content_length = response.headers.get("Content-Length")
+
+    if verbose and content_length is not None:
+        file_size = int(content_length)
+        num_bars = file_size // chunk_size
+        iterator = tqdm(iterator, total=num_bars, unit="KB", desc=url[:50], leave=True)
+
+    with open(file_path, "wb") as fp:
+        for chunk in iterator:
+            fp.write(chunk)
+
+    return response
 
 
 class HTTP(Service):
